+++ src/jsMain/kotlin/mu/two/KotlinLoggingLevel.kt
@@ -1,5 +1,5 @@
 package mu.two
 
-import mu.KotlinLoggingConfiguration.LOG_LEVEL
+import mu.two.KotlinLoggingConfiguration.LOG_LEVEL
 
 public fun mu.two.Level.isLoggingEnabled(): Boolean = this.ordinal >= LOG_LEVEL.ordinal
