+++ src/jvmMain/kotlin/mu/two/internal/LocationAwareKLogger.kt
@@ -3,8 +3,8 @@ package mu.two.internal
 import mu.two.KLogger
 import mu.two.KMarkerFactory
 import mu.two.Marker
-import mu.slf4j.Slf4jLoggerWrapper
-import mu.slf4j.toSlf4j
+import mu.two.slf4j.Slf4jLoggerWrapper
+import mu.two.slf4j.toSlf4j
 import org.slf4j.helpers.MessageFormatter
 import org.slf4j.spi.LocationAwareLogger
